generator/requirements.txt
jinja2
PyYAML
requests
